modal module refactor
